requirements.txt
asn1crypto==0.24.0
blinker==1.4
certifi==2018.11.29
cffi==1.11.5
Click==7.0
cryptography==2.4.2
Flask==1.0.2
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
peewee==3.7.1
psycopg2==2.7.6.1
pycparser==2.19
python-dotenv==0.10.1
python-telegram-bot==11.1.0
six==1.12.0
Werkzeug==0.14.1
